+++ core/server/api_container/server/service_network/partition_topology/partition_topology.go
@@ -6,7 +6,7 @@
 package partition_topology
 
 import (
-	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/user_service_registration"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/service"
 	"github.com/kurtosis-tech/kurtosis-core/server/api_container/server/service_network/service_network_types"
 	"github.com/kurtosis-tech/stacktrace"
 	"strings"
@@ -16,24 +16,25 @@ type PartitionConnection struct {
 	PacketLossPercentage float32
 }
 
+// TODO Add a RWMutex to this class to keep it threadsafe: https://github.com/kurtosis-tech/kurtosis-core/issues/582https://github.com/kurtosis-tech/kurtosis-core/issues/582
 // Stores the partition topology of the network, and exposes an API for modifying it
 type PartitionTopology struct {
 	defaultConnection PartitionConnection
 
-	servicePartitions map[user_service_registration.ServiceID]service_network_types.PartitionID
+	servicePartitions map[service.ServiceID]service_network_types.PartitionID
 
 	partitionConnections map[service_network_types.PartitionConnectionID]PartitionConnection
 
 	// A service can be a part of exactly one partition at a time
-	partitionServices map[service_network_types.PartitionID]map[user_service_registration.ServiceID]bool // partitionId -> set<serviceId>
+	partitionServices map[service_network_types.PartitionID]map[service.ServiceID]bool // partitionId -> set<serviceId>
 }
 
 func NewPartitionTopology(defaultPartition service_network_types.PartitionID, defaultConnection PartitionConnection) *PartitionTopology {
 	return &PartitionTopology{
-		servicePartitions:    map[user_service_registration.ServiceID]service_network_types.PartitionID{},
+		servicePartitions:    map[service.ServiceID]service_network_types.PartitionID{},
 		partitionConnections: map[service_network_types.PartitionConnectionID]PartitionConnection{},
-		partitionServices: map[service_network_types.PartitionID]map[user_service_registration.ServiceID]bool{
-			defaultPartition: map[user_service_registration.ServiceID]bool{},
+		partitionServices: map[service_network_types.PartitionID]map[service.ServiceID]bool{
+			defaultPartition: map[service.ServiceID]bool{},
 		},
 		defaultConnection: defaultConnection,
 	}
@@ -43,7 +44,7 @@ func NewPartitionTopology(defaultPartition service_network_types.PartitionID, de
 //                                        Public Methods
 // ================================================================================================
 func (topology *PartitionTopology) Repartition(
-	newPartitionServices map[service_network_types.PartitionID]map[user_service_registration.ServiceID]bool,
+	newPartitionServices map[service_network_types.PartitionID]map[service.ServiceID]bool,
 	newPartitionConnections map[service_network_types.PartitionConnectionID]PartitionConnection,
 	newDefaultConnection PartitionConnection) error {
 	// Validate we have at least one partition
@@ -52,15 +53,15 @@ func (topology *PartitionTopology) Repartition(
 	}
 
 	// Validate that each existing service in the testnet gets exactly one partition allocation
-	allServicesInNetwork := map[user_service_registration.ServiceID]bool{}
-	servicesNeedingAllocation := map[user_service_registration.ServiceID]bool{}
+	allServicesInNetwork := map[service.ServiceID]bool{}
+	servicesNeedingAllocation := map[service.ServiceID]bool{}
 	for serviceId := range topology.servicePartitions {
 		allServicesInNetwork[serviceId] = true
 		servicesNeedingAllocation[serviceId] = true
 	}
-	allocatedServices := map[user_service_registration.ServiceID]bool{}
-	unknownServices := map[user_service_registration.ServiceID]bool{}
-	duplicatedAllocations := map[user_service_registration.ServiceID]bool{}
+	allocatedServices := map[service.ServiceID]bool{}
+	unknownServices := map[service.ServiceID]bool{}
+	duplicatedAllocations := map[service.ServiceID]bool{}
 	for _, servicesForPartition := range newPartitionServices {
 		for serviceId := range servicesForPartition {
 			if doesServiceSetContainsElement(allocatedServices, serviceId) {
@@ -113,8 +114,8 @@ func (topology *PartitionTopology) Repartition(
 	}
 
 	// Defensive copies
-	newPartitionServicesCopy := map[service_network_types.PartitionID]map[user_service_registration.ServiceID]bool{}
-	newServicePartitionsCopy := map[user_service_registration.ServiceID]service_network_types.PartitionID{}
+	newPartitionServicesCopy := map[service_network_types.PartitionID]map[service.ServiceID]bool{}
+	newServicePartitionsCopy := map[service.ServiceID]service_network_types.PartitionID{}
 	for partitionId, servicesForPartition := range newPartitionServices {
 		newPartitionServicesCopy[partitionId] = copyServiceSet(servicesForPartition)
 		for serviceId := range servicesForPartition {
@@ -134,7 +135,7 @@ func (topology *PartitionTopology) Repartition(
 	return nil
 }
 
-func (topology *PartitionTopology) AddService(serviceId user_service_registration.ServiceID, partitionId service_network_types.PartitionID) error {
+func (topology *PartitionTopology) AddService(serviceId service.ServiceID, partitionId service_network_types.PartitionID) error {
 	if existingPartition, found := topology.servicePartitions[serviceId]; found {
 		return stacktrace.NewError(
 			"Cannot add service '%v' to partition '%v' because the service is already assigned to partition '%v'",
@@ -158,7 +159,7 @@ func (topology *PartitionTopology) AddService(serviceId user_service_registratio
 /*
 Removes the given service from the toplogy, if it exists. If it doesn't exist, this is a no-op.
 */
-func (topology *PartitionTopology) RemoveService(serviceId user_service_registration.ServiceID) {
+func (topology *PartitionTopology) RemoveService(serviceId service.ServiceID) {
 	partitionId, found := topology.servicePartitions[serviceId]
 	if !found {
 		return
@@ -169,15 +170,15 @@ func (topology *PartitionTopology) RemoveService(serviceId user_service_registra
 	delete(servicesForPartition, serviceId)
 }
 
-func (topology PartitionTopology) GetPartitionServices() map[service_network_types.PartitionID]map[user_service_registration.ServiceID]bool {
+func (topology PartitionTopology) GetPartitionServices() map[service_network_types.PartitionID]map[service.ServiceID]bool {
 	return topology.partitionServices
 }
 
-func (topology PartitionTopology) GetServicePacketLossConfigurationsByServiceID() (map[user_service_registration.ServiceID]map[user_service_registration.ServiceID]float32, error) {
-	result := map[user_service_registration.ServiceID]map[user_service_registration.ServiceID]float32{}
+func (topology PartitionTopology) GetServicePacketLossConfigurationsByServiceID() (map[service.ServiceID]map[service.ServiceID]float32, error) {
+	result := map[service.ServiceID]map[service.ServiceID]float32{}
 	for partitionId, servicesInPartition := range topology.partitionServices {
 		for serviceId := range servicesInPartition {
-			otherServicesPacketLossConfigMap := map[user_service_registration.ServiceID]float32{}
+			otherServicesPacketLossConfigMap := map[service.ServiceID]float32{}
 			for otherPartitionId, servicesInOtherPartition := range topology.partitionServices {
 				if partitionId == otherPartitionId {
 					// Two services in the same partition will never block each other
@@ -217,7 +218,7 @@ func (topology PartitionTopology) getPartitionConnection(
 	return connection, nil
 }
 
-func serviceIdSetToCommaStr(serviceSet map[user_service_registration.ServiceID]bool) string {
+func serviceIdSetToCommaStr(serviceSet map[service.ServiceID]bool) string {
 	strSlice := []string{}
 	for serviceId := range serviceSet {
 		strSlice = append(strSlice, string(serviceId))
@@ -225,15 +226,15 @@ func serviceIdSetToCommaStr(serviceSet map[user_service_registration.ServiceID]b
 	return strings.Join(strSlice, ", ")
 }
 
-func doesServiceSetContainsElement(serviceSet map[user_service_registration.ServiceID]bool, element user_service_registration.ServiceID) bool {
+func doesServiceSetContainsElement(serviceSet map[service.ServiceID]bool, element service.ServiceID) bool {
 	if _, found := serviceSet[element]; found {
 		return true
 	}
 	return false
 }
 
-func copyServiceSet(serviceSet map[user_service_registration.ServiceID]bool) map[user_service_registration.ServiceID]bool {
-	newServiceSet := map[user_service_registration.ServiceID]bool{}
+func copyServiceSet(serviceSet map[service.ServiceID]bool) map[service.ServiceID]bool {
+	newServiceSet := map[service.ServiceID]bool{}
 	for serviceGuid := range serviceSet {
 		newServiceSet[serviceGuid] = true
 	}