+++ core/server/api_container/server/service_network/service_network_test.go
@@ -7,8 +7,9 @@ package service_network
 
 import (
 	"context"
+	"fmt"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/enclave"
-	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/user_service_registration"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/service"
 	"github.com/kurtosis-tech/kurtosis-core/server/api_container/server/service_network/networking_sidecar"
 	"github.com/stretchr/testify/require"
 	"net"
@@ -18,41 +19,40 @@ import (
 
 const (
 	packetLossConfigForBlockedPartition = float32(100)
+
+	numServices = 10
 )
 
 func TestUpdateTrafficControl(t *testing.T) {
-	numServices := 10
 	ctx := context.Background()
 
 	enclaveId := enclave.EnclaveID("test")
 
-	sidecars := map[user_service_registration.ServiceID]networking_sidecar.NetworkingSidecarWrapper{}
-	mockSidecars := map[user_service_registration.ServiceID]*networking_sidecar.MockNetworkingSidecarWrapper{}
+	sidecars := map[service.ServiceID]networking_sidecar.NetworkingSidecarWrapper{}
+	registrations := map[service.ServiceID]*service.ServiceRegistration{}
+	mockSidecars := map[service.ServiceID]*networking_sidecar.MockNetworkingSidecarWrapper{}
 	for i := 0; i < numServices; i++ {
 		serviceId := testServiceIdFromInt(i)
+
 		sidecar := networking_sidecar.NewMockNetworkingSidecarWrapper()
 		sidecars[serviceId] = sidecar
 		mockSidecars[serviceId] = sidecar
-	}
 
-	serviceRegistrationByServiceId := map[user_service_registration.ServiceID]*user_service_registration.UserServiceRegistration{}
-	for i := 0; i < numServices; i++ {
-		registrationGuid := testServiceRegistrationGuidFromInt(i)
-		serviceId := testServiceIdFromInt(i)
 		ip := testIpFromInt(i)
-		serviceRegistrationByServiceId[serviceId] = user_service_registration.NewUserServiceRegistration(
-			registrationGuid,
-			enclaveId,
+		serviceGuid := testServiceGuidFromInt(i)
+		registrations[serviceId] = service.NewServiceRegistration(
 			serviceId,
+			serviceGuid,
+			enclaveId,
 			ip,
 		)
 	}
 
 	// Creates the pathological "line" of connections, where each service can only see the services adjacent
-	targetServicePacketLossConfigs := map[user_service_registration.ServiceID]map[user_service_registration.ServiceID]float32{}
+	targetServicePacketLossConfigs := map[service.ServiceID]map[service.ServiceID]float32{}
 	for i := 0; i < numServices; i++ {
 		serviceId := testServiceIdFromInt(i)
-		otherServicesPacketLossConfig := map[user_service_registration.ServiceID]float32{}
+		otherServicesPacketLossConfig := map[service.ServiceID]float32{}
 		for j := 0; j < numServices; j++ {
 			if j < i-1 || j > i+1 {
 				blockedServiceId := testServiceIdFromInt(j)
@@ -62,7 +62,7 @@ func TestUpdateTrafficControl(t *testing.T) {
 		targetServicePacketLossConfigs[serviceId] = otherServicesPacketLossConfig
 	}
 
-	require.Nil(t, updateTrafficControlConfiguration(ctx, targetServicePacketLossConfigs, serviceRegistrationByServiceId, sidecars))
+	require.Nil(t, updateTrafficControlConfiguration(ctx, targetServicePacketLossConfigs, registrations, sidecars))
 
 	// Verify that each service got told to block exactly the right things
 	for i := 0; i < numServices; i++ {
@@ -90,10 +90,14 @@ func testIpFromInt(i int) net.IP {
 	return []byte{1, 1, 1, byte(i)}
 }
 
-func testServiceIdFromInt(i int) user_service_registration.ServiceID {
-	return user_service_registration.ServiceID("service-" + strconv.Itoa(i))
+func testServiceIdFromInt(i int) service.ServiceID {
+	return service.ServiceID("service-" + strconv.Itoa(i))
 }
 
-func testServiceRegistrationGuidFromInt(i int) user_service_registration.UserServiceRegistrationGUID {
-	return user_service_registration.UserServiceRegistrationGUID("registration-" + strconv.Itoa(i))
+func testServiceGuidFromInt(i int) service.ServiceGUID {
+	return service.ServiceGUID(fmt.Sprintf(
+		"service-%v-%v",
+		i,
+		i,
+	))
 }